requirements.txt
pandas
plotly_express
streamlit==1.12.2
altair<5